OLS-UI rebuild Sidebar turn into Navbar
