+++ OLS-UI/ols-ui/src/layouts/DefaultLayout/DefaultLayout.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 // components
 import styles from './DefaultLayout.module.scss';
 import Header from '~/layouts/components/Header';
-import Sidebar from '~/layouts/components/Sidebar';
+import Navbar from '~/layouts/components/Navbar';
 
 const cx = classNames.bind(styles);
 
@@ -14,8 +14,8 @@ const DefaultLayout = ({ children }) => {
     return (
         <div className={cx('wrapper')}>
             <Header />
+            <Navbar />
             <div className={cx('container')}>
-                <Sidebar />
                 <div className={cx('content')}>{children}</div>
             </div>
         </div>
