+++ OLS-UI/ols-ui/src/layouts/DefaultLayout/DefaultLayout.module.scss
@@ -1,14 +1,19 @@
 .wrapper {
     display: flex;
     flex-direction: column;
-    align-items: center;
+    //align-items: center;
+
+    // new
+    align-items: flex-start;
+    padding-left: 29px;
 }
 
 .container {
     width: var(--default-layout-header-width);
     max-width: 100%;
     padding: 0 var(--default-layout-header-horizontal-spacer);
-    margin-top: var(--default-layout-header-height, 60px);
+    //margin-top: var(--default-layout-header-height, 60px);
+    margin-top: 130px;
     display: flex;
 }
