+++ OLS-UI/ols-ui/src/layouts/components/Header/Header.module.scss
@@ -6,18 +6,18 @@
     --search-button-width: 52px;
 
     // sap xep 
-    // posistion
+    // posistion - vi tri
     position: fixed;
     top: 0;
     left: 0;
     z-index: 10; // set khoảng - chưa hiểu lắm 
-    // display
+    // display - trinh bay
     display: flex;
     justify-content: center;
-    // box model 
+    // box model - cac model boxes
     width: 100%;
     height: var(--default-layout-header-height, 60px);
-    box-shadow: 0px 1px 1px rgb(0 0 0 / 12%);
+    //box-shadow: 0px 1px 1px rgb(0 0 0 / 12%);
     background-color: var(--white);
 }
