+++ OLS-UI/ols-ui/src/layouts/components/Navbar/Menu/Menu.module.scss
@@ -1,8 +1,9 @@
 .menu-item {
-    display: flex;
+    display: inline-flex; // mới
+    //display: flex; // cũ
     align-items: center;
-    padding: 8px;
-    font-size: 1.2rem;
+    padding: 8px 18px 8px 8px;
+    font-size: 0.9rem;
     font-weight: 500; // độ dày
     border-radius: 5px;
     color: var(--text-color);
