+++ OLS-UI/ols-ui/src/layouts/components/Navbar/Navbar.module.scss
@@ -1,0 +1,14 @@
+.wrapper {
+    // position -vi tri
+    position: fixed;
+    left: 0;
+    padding-left: 1.2rem;
+    padding-top: 57px;
+    // display - trinh bay 
+    // boxes models
+    width: 100%;
+    height: 100px;
+    // colors
+    background-color: var(--white);
+    box-shadow: 0px 1px 1px rgb(0 0 0 / 12%);
+}
