--- OLS-UI/ols-ui/src/layouts/components/Sidebar/Sidebar.module.scss
@@ -1,8 +1,0 @@
-.wrapper {
-    width: 356px;
-    padding-right: 20px;
-    padding-left: 0px;
-    padding-top: 26px;
-    padding-bottom: 8px;
-    //padding: 20px 0px 26px 8px;
-}
