--- OLS-UI/ols-ui/src/layouts/components/Sidebar/index.js
@@ -1,1 +1,0 @@
-export { default } from './Sidebar';
