+++ OLS-UI/ols-ui/src/pages/Home/Home.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Home = () => {
+    return (
+        // đang hardcode
+        <div>
+            <p style={{ height: 2000 }}>Home</p>
+        </div>
+    );
+};
+
+export default Home;
