+++ OLS-UI/ols-ui/src/pages/Home/index.js
@@ -1,11 +1,1 @@
-import React from 'react';
-
-const Home = () => {
-    return (
-        <div>
-            <p>Home</p>
-        </div>
-    );
-};
-
-export default Home;
+export { default } from './Home';
