+++ OLS-UI/ols-ui/src/pages/LearningPaths/LearningPaths.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const LearningPaths = () => {
+    return (
+        // đang hardcode
+        <div style={{ height: 2000 }}>LearningPaths</div>
+    );
+};
+
+export default LearningPaths;
