+++ OLS-UI/ols-ui/src/pages/LearningPaths/index.js
@@ -1,7 +1,1 @@
-import React from 'react';
-
-const LearningPaths = () => {
-    return <div>LearningPaths</div>;
-};
-
-export default LearningPaths;
+export { default } from './LearningPaths';
