+++ OLS-UI/ols-ui/src/pages/Search/Search.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Search = () => {
+    return (
+        // đang hardcode
+        <div style={{ height: 2000 }}>Search</div>
+    );
+};
+
+export default Search;
