+++ OLS-UI/ols-ui/src/pages/Search/index.js
@@ -1,7 +1,1 @@
-import React from 'react';
-
-const Search = () => {
-    return <div>Search</div>;
-};
-
-export default Search;
+export { default } from './Search';
